Scripts - HRRR - Add index creation for wgrib files
Add the creation of an index file for each downloaded, cropped, and filtered wgrib file. This will help with file access when reading in a file and reduce seeking through each file upon read. While this is not a huge performance boost, it is a good practice to have and the list of used HRRR variables are growing.